[refactor] Use camelCase for authenticationKey
