+++ src/core/account.ts
@@ -1,8 +1,8 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { AccountAddress } from "./account_address";
-import { AuthenticationKey } from "./authentication_key";
+import { AccountAddress } from "./accountAddress";
+import { AuthenticationKey } from "./authenticationKey";
 import { PrivateKey, PublicKey, Signature } from "./crypto/asymmetricCrypto";
 import { Ed25519PrivateKey, Ed25519PublicKey } from "./crypto/ed25519";
 import { MultiEd25519PublicKey } from "./crypto/multiEd25519";
