+++ src/core/index.ts
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from "./account";
-export * from "./account_address";
-export * from "./authentication_key";
+export * from "./accountAddress";
+export * from "./authenticationKey";
 export * from "./common";
 export * from "./crypto";
 export * from "./hex";
